+++ .prettierrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  printWidth: 120,
-  semi: false,
-  singleQuote: true,
-  eslintIntegration: true,
-  stylelintIntegration: true,
-  bracketSpacing: true
-}