+++ register.js
@@ -1,2 +1,0 @@
-require('babel-register')
-const app = require('./index')